add a stickyHeader prop
