+++ packages/material-ui/src/Table/Table.d.ts
@@ -5,6 +5,7 @@ export interface TableProps extends StandardProps<TableBaseProps, TableClassKey>
   component?: React.ElementType<TableBaseProps>;
   padding?: Padding;
   size?: Size;
+  stickyHeader?: boolean;
 }
 
 export type TableBaseProps = React.TableHTMLAttributes<HTMLTableElement>;
@@ -13,7 +14,7 @@ export type Padding = 'default' | 'checkbox' | 'none';
 
 export type Size = 'small' | 'medium';
 
-export type TableClassKey = 'root';
+export type TableClassKey = 'root' | 'stickyHeader';
 
 declare const Table: React.ComponentType<TableProps>;
